requirements.txt
# Core Dependencies
langchain==0.2.0
langchain-core==0.3.83
langchain-community==0.2.0
langgraph==0.2.43
langchain-text-splitters==0.2.1
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.6.9

# Vector Database
pymilvus==2.4.0
langchain-milvus>=0.1.0

# LLM Providers
ollama==0.1.5

# Document Processing
docling==2.10.0
docling-core==2.64.0
docling-ibm-models==2.0.8
docling-parse==3.4.0
python-docx==1.1.2
pypdf==4.3.1
python-pptx==1.0.2
pandas==2.3.3
openpyxl==3.1.5
easyocr==1.7.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
lxml==5.4.0
beautifulsoup4==4.14.3

# Embeddings
sentence-transformers==3.0.1
transformers==4.40.0
torch==2.10.0
torchvision==0.25.0

# UI and Web
streamlit==1.32.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
rich==13.9.4
tqdm==4.67.3
click==8.3.1
typer==0.12.5
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3

# Search and Retrieval
rank-bm25==0.2.2
faiss-cpu==1.13.2

# Async and Concurrency
asyncio-throttle==1.0.2
aiofiles==23.2.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# Development
black==24.3.0
ruff==0.4.5
mypy==1.9.0

# Additional dependencies that may be needed
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.0
joblib==1.5.3
threadpoolctl==3.6.0
tokenizers==0.19.1
safetensors==0.7.0
filelock==3.20.3
huggingface-hub==0.36.2
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
requests==2.32.5